+++ TiendaRPG/src/tiendarpg/Consumible.java
@@ -14,13 +14,13 @@ public class Consumible extends Item {
     private String tipo; 
     
     //Constructor que modifica los atributos de Consumible
-    public Consumible(String nombre, String color, double precio, int nivel, int poder, String tipo){
+    public Consumible(String nombre, String color, int precio, int nivel, int poder, String tipo){
         
         super(nombre,color,precio,nivel,poder);
         this.tipo= tipo;
     }
 
-    //toString, muestra el contenido de los atributos de Arma
+    //toString, muestra el contenido de los atributos de Consumible
     @Override
     public String toString  (){ //Getter
         return super.getDatos()+" tipo: "+ tipo;
